feat: release notes details page
