+++ packages/gatsby-theme-docs/src/layouts/internals/layout-release-note.js
@@ -1,54 +1,14 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import { MDXRenderer } from 'gatsby-plugin-mdx';
-import styled from '@emotion/styled';
-import { ThemeProvider as UiKitThemeProvider } from 'emotion-theming';
 import SpacingsStack from '@commercetools-uikit/spacings-stack';
-import SpacingsInline from '@commercetools-uikit/spacings-inline';
-import Stamp from '@commercetools-uikit/stamp';
-import { Markdown, designSystem } from '@commercetools-docs/ui-kit';
-
-const DateElement = styled.div`
-  line-height: ${designSystem.typography.lineHeights.small};
-`;
-const Topics = styled.div`
-  color: ${designSystem.colors.light.textInfo};
-  font-size: ${designSystem.typography.fontSizes.small};
-
-  > * + * {
-    padding-left: ${designSystem.dimensions.spacings.xs};
-    margin-left: ${designSystem.dimensions.spacings.xs};
-    border-left: 1px solid ${designSystem.colors.light.surfaceSecondary3};
-  }
-`;
-const stampTheme = {
-  fontSizeDefault: designSystem.typography.fontSizes.extraSmall,
-  // Override the `critical` style which is used for the "fix" type
-  colorError95: designSystem.colors.light.surfaceForReleaseNoteTypeFix,
-  colorError: designSystem.colors.light.borderForReleaseNoteTypeFix,
-};
+import { Markdown } from '@commercetools-docs/ui-kit';
+import LayoutReleaseNoteBody from './layout-release-note-body';
 
 const ReleaseNote = (props) => {
   return (
     <SpacingsStack scale="m">
       <Markdown.H3>{props.title}</Markdown.H3>
-      <DateElement>{props.date}</DateElement>
-      <UiKitThemeProvider theme={stampTheme}>
-        <SpacingsInline>
-          <Stamp tone={mapTypeToTone(props)}>{mapTypeToLabel(props)}</Stamp>
-        </SpacingsInline>
-      </UiKitThemeProvider>
-      {props.topics.length > 0 && (
-        <Topics>
-          {props.topics.map((topic) => (
-            <span key={topic}>{topic}</span>
-          ))}
-        </Topics>
-      )}
-
-      <div>
-        <MDXRenderer>{props.body}</MDXRenderer>
-      </div>
+      <LayoutReleaseNoteBody {...props} />
     </SpacingsStack>
   );
 };
@@ -63,29 +23,3 @@ ReleaseNote.propTypes = {
 };
 
 export default ReleaseNote;
-
-function mapTypeToTone(props) {
-  switch (props.type) {
-    case 'feature':
-      return 'positive';
-    case 'enhancement':
-      return 'information';
-    case 'fix':
-      return 'critical';
-    default:
-      return props.type;
-  }
-}
-
-function mapTypeToLabel(props) {
-  switch (props.type) {
-    case 'feature':
-      return 'Feature';
-    case 'enhancement':
-      return 'Enhancement';
-    case 'fix':
-      return 'Resolved Issue';
-    default:
-      return props.type;
-  }
-}
